# docs/api.openapi.yaml
openapi: 3.1.0
info:
  title: Redaê API
  version: 1.0.0
  description: Contratos HTTP do MVP do Redaê.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /admin/users/count:
    get:
      summary: Consulta o total de usuários
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Error' }
  /admin/users:
    get:
      summary: Lista usuários e saldos de créditos para administradores
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: sort
          in: query
          schema: { type: string, example: 'createdAt,desc' }
        - name: search
          in: query
          schema: { type: string, maxLength: 120 }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Error' }
  /admin/credit-adjustments:
    post:
      summary: Concede creditos a um usuario
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminCreditAdjustmentRequest' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    AdminCreditAdjustmentRequest:
      type: object
      required: [userId, credits]
      properties:
        userId: { type: string, format: uuid }
        credits: { type: integer, minimum: 1 }
  responses:
    Success:
      description: Operação concluída
      content:
        application/json:
          schema: { type: object, required: [data, meta, traceId] }
    Error:
      description: Erro da API
      content:
        application/json:
          schema:
            type: object
            required: [error, traceId]
            properties:
              error: { type: object }
              traceId: { type: string }